# [libc++] Optimize lexicographical_compare