--- a/libcxx/test/benchmarks/CMakeLists.txt
+++ b/libcxx/test/benchmarks/CMakeLists.txt
@@ -114,6 +114,7 @@ set(BENCHMARK_TESTS
     algorithms/find.bench.cpp
     algorithms/fill.bench.cpp
     algorithms/for_each.bench.cpp
+    algorithms/lexicographical_compare.bench.cpp
     algorithms/lower_bound.bench.cpp
     algorithms/make_heap.bench.cpp
     algorithms/make_heap_then_sort_heap.bench.cpp