--- a/libcxx/test/libcxx/transitive_includes/cxx14.csv
+++ b/libcxx/test/libcxx/transitive_includes/cxx14.csv
@@ -41,7 +41,9 @@ array algorithm
 array compare
 array concepts
 array cstddef
+array cstdint
 array cstdlib
+array cwchar
 array initializer_list
 array iterator
 array limits
@@ -295,6 +297,7 @@ forward_list cstddef
 forward_list cstdint
 forward_list cstdlib
 forward_list cstring
+forward_list cwchar
 forward_list functional
 forward_list initializer_list
 forward_list iosfwd
@@ -455,6 +458,7 @@ list cstddef
 list cstdint
 list cstdlib
 list cstring
+list cwchar
 list functional
 list initializer_list
 list iosfwd
@@ -496,7 +500,9 @@ locale version
 map compare
 map concepts
 map cstddef
+map cstdint
 map cstdlib
+map cwchar
 map functional
 map initializer_list
 map iterator
@@ -732,7 +738,9 @@ semaphore version
 set compare
 set concepts
 set cstddef
+set cstdint
 set cstdlib
+set cwchar
 set functional
 set initializer_list
 set iterator