@EricWF 2023-09-05
Can you please add benchmarks for types other than `int` or other small integral types.
Making an algorithm faster for types that fit in registers and are cheap to copy isn't the whole story, and might actually hide the pessimization for other types.


std::string is a good candidate for a test type. With both short and long strings.