@EricWF 2023-09-05
Please add a description of *how* you optimized the code. It's not at all clear after reading the diff.