@EricWF 2023-09-05
Please take a look at this godbolt.

It looks like under optimized builds, apart from for `unsigned char`, there is literally no difference in generated assembly.
However, for non-optimized builds, the new implementation is like 3x the size of the old one.

I don't think that cost/benefit trade off has the correct balance. 

Convince me otherwise?