@EricWF 2023-09-05
@philnik777 Also, please include descriptions. IT took me an hour to realize this change only affects `unsigned char`.