@EricWF 2023-09-07
Drafts seem to prevent merging, not review.

Is there a way to filter pull requests to non-draft libc++ PR's? If not, then I think we need to reconsider the purpose of drafts. 

What is the purpose of opening the draft if you didn't want review?