@philnik777 2023-09-07
> Drafts seem to prevent merging, not review.

Yes. It also prevents notifying people until it is made a PR.

> Is there a way to filter pull requests to non-draft libc++ PR's? If not, then I think we need to reconsider the purpose of drafts.

[`-is:draft label:libc++`](https://github.com/llvm/llvm-project/pulls?q=is%3Apr+is%3Aopen+label%3Alibc%2B%2B+-is%3Adraft)

> What is the purpose of opening the draft if you didn't want review?

It has multiple benefits:
- People see what I'm working on and don't duplicate it
- I can work on any comments etc. for the PR
- I can easily view the diff and make sure that the patch applies properly
- The CI is run, so I can make sure that I'm not missing something super obvious
- It allows people to give high-level feedback (i.e. I can upload a diff and bounce ideas off people)